# openapi.yaml
openapi: 3.1.0
info:
  title: OpenFamHub API
  version: "1.0.0"
  description: |
    Read-mostly aggregation API for the OpenFamHub family command center. Aggregates
    external sources (Todoist, iCal, Mealie, a Monarch Money sidecar, Open-Meteo)
    behind one small backend so credentials never reach the browser, plus a
    small first-party economy (chore points -> reward redemption).

    Almost every read endpoint returns **503** when its source isn't configured
    (missing token/URL/etc.) rather than an error — the frontend keeps rendering
    from bundled demo data in that case. See `GET /api/health` for a
    per-source configured/not-configured summary.

    Data-source endpoints (tasks, calendar, meals, budget, weather, photos) are
    served by whichever provider is selected in `server/providers/registry.js`;
    paths/shapes below are the *contract* the active provider must satisfy, not
    tied to any one vendor.
  license:
    name: MIT
servers:
  - url: /api
...
paths:
...
  /settings:
    get:
      tags: [meta]
      summary: Current settings (parent-gated — same unlock as Budget)
      description: >-
        401 until a parent PIN unlock (POST /budget/unlock) — the Settings
        screen uses that to show its PIN gate.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [meta]
      summary: Update settings (parent-gated)
      description: >-
        Partial update; every provided key must be whitelisted and in range or
        the whole request 400s (all-or-nothing). Publishes a 'settings' event
        on /live so every connected wall re-applies immediately.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: The full resolved settings after the update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /presence:
    get:
      tags: [meta]
      summary: Last reported presence state
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  present: { type: boolean }
                  changedAt: { type: string, format: date-time, nullable: true }
    post:
      tags: [meta]
      summary: Report presence (wall sensor scaffold)
      description: >-
        For a future PIR/mmWave sensor (ESP32 or anything that can HTTP POST).
        A change publishes `{"kind":"presence"}` on /live; the wall wakes its
        screensaver (deliberately not scheduled sleep). If PRESENCE_TOKEN is
        set, the request must carry it in an X-Presence-Token header.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [present]
              properties:
                present: { type: boolean }
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /tasks/{list}:
    get:
      tags: [tasks]
      summary: A flat task list for a dashboard card (currently just `todos`)
      parameters:
        - $ref: "#/components/parameters/ListName"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskList" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/NotConfigured" }
  /tasks/{id}/{action}:
    post:
      tags: [tasks]
      summary: Close or reopen a Todoist task (tap-to-complete)
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/CloseReopenAction"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /board/{board}:
    get:
      tags: [tasks]
      summary: Per-member task board (`todo` | `routine`), split by assignee
      parameters:
        - name: board
          in: path
          required: true
          schema: { type: string, enum: [todo, routine] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Board" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/NotConfigured" }
  /board/routine/{id}/{action}:
    post:
      tags: [tasks]
      summary: Close/reopen a Routine·Chore task and award/revert its point
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/CloseReopenAction"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [memberId]
              properties:
                memberId: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  balance: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /todoist/collaborators:
    get:
      tags: [tasks]
      summary: Discovery helper — list collaborators of the configured board projects
      description: Used once, at setup time, to find each member's Todoist UID for `.env`.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      type: object
                      properties:
                        project: { type: string }
                        collaborators:
                          type: array
                          items:
                            type: object
                            properties:
                              id: { type: string }
                              name: { type: string }
                              email: { type: string }
        "503": { $ref: "#/components/responses/NotConfigured" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: openfamhub_session
  parameters:
    ListName:
      name: list
      in: path
      required: true
      schema: { type: string, enum: [todos] }
    TaskId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    CloseReopenAction:
      name: action
      in: path
      required: true
      schema: { type: string, enum: [close, reopen] }
...
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean } }
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Unknown resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Not logged in
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotConfigured:
      description: Source not configured, or upstream unreachable — client falls back to demo data
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UpstreamError:
      description: Upstream (e.g. Todoist) rejected the write
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Settings:
      type: object
      description: >-
        The wall-editable settings (server/settings/index.js is the
        whitelist). Resolution per key — DB row (saved on the Settings
        screen) → .env → hardcoded default.
      properties:
        sleepStartHour: { type: integer, minimum: 0, maximum: 23 }
        sleepEndHour: { type: integer, minimum: 0, maximum: 23 }
        screensaverIdleMin: { type: integer, minimum: 1, maximum: 120 }
        screensaverCycleSec: { type: integer, minimum: 5, maximum: 300 }
        pushChoreReminder: { type: integer, enum: [0, 1], description: "Kill-switch for the scheduled morning chore reminder" }
        pushDinnerDigest: { type: integer, enum: [0, 1], description: "Kill-switch for the scheduled dinner digest" }
...
    Error:
      type: object
      properties:
        error: { type: string }
...
    Task:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        done: { type: boolean }
    TaskList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Task" }
        count: { type: string }
        left: { type: string }
    Board:
      type: object
      properties:
        members:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              tasks:
                type: array
                items: { $ref: "#/components/schemas/Task" }
...